feat(SignMessageOnChain): implement confirmation flow and details for signing messages on-chain
