+++ apps/web/src/components/tx-flow/flows/SignMessageOnChain/ReviewSignMessageOnChain.test.tsx
@@ -105,6 +105,7 @@ describe('ReviewSignMessageOnChain', () => {
               },
             }}
             method={Methods.signTypedMessage}
+            onSubmit={() => {}}
           />
         </SafeTxContext.Provider>,
       )
